スキャン関連のロジックを修正
